test : 기능 단위 클래스 상세 분리로 인해 테스트 클래스도 동일하게 적용
- ProfileServiceTest, HistoryServiceTest로 클래스 분리

ref : #224
