+++ src/test/java/com/example/baseballprediction/domain/member/service/ProfileServiceTest.java
@@ -9,13 +9,11 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.data.domain.Page;
 import org.springframework.test.context.ActiveProfiles;
 
 import com.example.baseballprediction.domain.chat.dto.ChatRequestDTO;
 import com.example.baseballprediction.domain.gifttoken.entity.GiftToken;
 import com.example.baseballprediction.domain.gifttoken.repository.GiftTokenRepository;
-import com.example.baseballprediction.domain.member.dto.FairyProjection;
 import com.example.baseballprediction.domain.member.dto.MemberRequest;
 import com.example.baseballprediction.domain.member.dto.MemberResponse;
 import com.example.baseballprediction.domain.member.dto.ProfileProjection;
@@ -32,56 +30,24 @@ import com.example.baseballprediction.global.error.exception.BusinessException;
 
 @ActiveProfiles("test")
 @SpringBootTest
-class MemberServiceTest {
+public class ProfileServiceTest {
+	@Autowired
+	private MemberRepository memberRepository;
 
 	@Autowired
-	MemberService memberService;
+	private TeamRepository teamRepository;
 
 	@Autowired
-	MemberRepository memberRepository;
+	private MonthlyFairyRepository monthlyFairyRepository;
 
 	@Autowired
-	TeamRepository teamRepository;
+	private GiftTokenRepository giftTokenRepository;
 
 	@Autowired
-	MonthlyFairyRepository monthlyFairyRepository;
+	private ProfileService profileService;
 
 	@Autowired
-	GiftTokenRepository giftTokenRepository;
-
-	@Autowired
-	MemberServiceHelper memberServiceHelper;
-
-	// @BeforeEach
-	// void setUp() {
-	// 	Team team1 = Team.builder()
-	// 		.name("LG 트윈스")
-	// 		.shortName("LG")
-	// 		.build();
-	//
-	// 	Team team2 = Team.builder()
-	// 		.name("한화 이글스")
-	// 		.shortName("한화")
-	// 		.build();
-	//
-	// 	teamRepository.saveAll(List.of(team1, team2));
-	//
-	// 	Member member1 = Member.builder()
-	// 		.socialType(SocialType.KAKAO)
-	// 		.nickname("테스트유저1")
-	// 		.username("playdot1")
-	// 		.password("123")
-	// 		.build();
-	//
-	// 	Member member2 = Member.builder()
-	// 		.socialType(SocialType.KAKAO)
-	// 		.nickname("테스트유저2")
-	// 		.username("playdot2")
-	// 		.password("123")
-	// 		.build();
-	//
-	// 	memberRepository.saveAll(List.of(member1, member2));
-	// }
+	private MemberServiceHelper memberServiceHelper;
 
 	@AfterEach
 	void tearDown() {
@@ -112,7 +78,7 @@ class MemberServiceTest {
 		Team savedFavoriteTeam = teamRepository.save(favoriteTeam);
 
 		//when
-		memberService.modifyLikeTeam(savedMember.getUsername(), savedFavoriteTeam.getId());
+		profileService.modifyLikeTeam(savedMember.getUsername(), savedFavoriteTeam.getId());
 
 		//then
 		Team updateFavoriteTeam = memberServiceHelper.getFavoriteTeam(savedMember.getId());
@@ -137,7 +103,7 @@ class MemberServiceTest {
 
 		MemberRequest.DetailsDTO detailsDTO = new MemberRequest.DetailsDTO("테스트유저2", "한마디");
 		//when
-		memberService.modifyDetails("playdot1", detailsDTO, null);
+		profileService.modifyDetails("playdot1", detailsDTO, null);
 
 		//then
 		Member updateMember = memberRepository.findByUsername("playdot1").orElseThrow();
@@ -170,9 +136,9 @@ class MemberServiceTest {
 		Member updateMember = memberRepository.findByUsername("playdot1").orElseThrow();
 
 		//when
-		MemberResponse.NicknameDTO trueResult = memberService.findExistNickname(updateMember.getId(),
+		MemberResponse.NicknameDTO trueResult = profileService.findExistNickname(updateMember.getId(),
 			existNickname);
-		MemberResponse.NicknameDTO falseResult = memberService.findExistNickname(updateMember.getId(),
+		MemberResponse.NicknameDTO falseResult = profileService.findExistNickname(updateMember.getId(),
 			newNickname);
 
 		//then
@@ -195,7 +161,7 @@ class MemberServiceTest {
 		Member savedMember = memberRepository.save(member);
 
 		//when
-		ProfileProjection result = memberService.findProfile(savedMember.getNickname());
+		ProfileProjection result = profileService.findProfile(savedMember.getNickname());
 
 		//then
 		assertThat(result)
@@ -244,7 +210,7 @@ class MemberServiceTest {
 		monthlyFairyRepository.saveAll(List.of(win1, win2, lose1));
 
 		//when
-		ProfileProjection result = memberService.findProfile(savedMember.getNickname());
+		ProfileProjection result = profileService.findProfile(savedMember.getNickname());
 
 		//then
 		assertThat(result)
@@ -266,163 +232,11 @@ class MemberServiceTest {
 		Member savedMember = memberRepository.save(member);
 
 		//when
-		MemberResponse.ProfileDTO result = memberService.findDetails(savedMember.getUsername());
+		MemberResponse.ProfileDTO result = profileService.findDetails(savedMember.getUsername());
 
 		//then
 		assertThat(result).extracting("nickname", "profileImageUrl", "level", "token", "teamName", "comment")
 			.contains("테스트유저1", null, 1, 0);
-	}
-
-	@DisplayName("승리요정/패배요정 정보를 조회한다.")
-	@Test
-	void findFairyStatistics() {
-		//given
-		Member member = Member.builder()
-			.username("playdot1")
-			.password("asd")
-			.nickname("테스트유저1")
-			.socialType(SocialType.KAKAO)
-			.build();
-
-		Member savedMember = memberRepository.save(member);
-
-		MonthlyFairy win1 = MonthlyFairy.builder()
-			.type(FairyType.WIN)
-			.month(202403)
-			.member(savedMember)
-			.rank(1)
-			.voteRatio(70)
-			.build();
-
-		MonthlyFairy win2 = MonthlyFairy.builder()
-			.type(FairyType.WIN)
-			.month(202402)
-			.member(savedMember)
-			.rank(2)
-			.voteRatio(60)
-			.build();
-
-		MonthlyFairy lose1 = MonthlyFairy.builder()
-			.type(FairyType.LOSE)
-			.month(202402)
-			.member(savedMember)
-			.rank(1)
-			.voteRatio(5)
-			.build();
-
-		monthlyFairyRepository.saveAll(List.of(win1, win2, lose1));
-
-		//when
-		List<FairyProjection> result = memberService.findFairyStatistics(savedMember.getId());
-
-		//then
-		assertThat(result).hasSize(3)
-			.extracting("title", "count")
-			.containsExactlyInAnyOrder(
-				tuple("승리요정 1등", 1),
-				tuple("승리요정 2등", 1),
-				tuple("패배요정 1등", 1)
-			);
-	}
-
-	@DisplayName("월간 승리요정 히스토리를 조회한다.")
-	@Test
-	void findFairyHistories() {
-		//given
-		Member member = Member.builder()
-			.nickname("테스트유저1")
-			.username("playdot1")
-			.password("123")
-			.socialType(SocialType.KAKAO)
-			.build();
-
-		Member savedMember = memberRepository.save(member);
-
-		MonthlyFairy win1 = MonthlyFairy.builder()
-			.type(FairyType.WIN)
-			.month(202403)
-			.member(savedMember)
-			.rank(1)
-			.voteRatio(70)
-			.build();
-
-		MonthlyFairy win2 = MonthlyFairy.builder()
-			.type(FairyType.WIN)
-			.month(202402)
-			.member(savedMember)
-			.rank(2)
-			.voteRatio(60)
-			.build();
-
-		MonthlyFairy lose1 = MonthlyFairy.builder()
-			.type(FairyType.LOSE)
-			.month(202402)
-			.member(savedMember)
-			.rank(1)
-			.voteRatio(5)
-			.build();
-
-		monthlyFairyRepository.saveAll(List.of(win1, win2, lose1));
-
-		//when
-		Page<MemberResponse.FairyHistoryDTO> result = memberService.findFairyHistories(savedMember.getId(), 0,
-			15);
-
-		//then
-		assertThat(result.get()).hasSize(3)
-			.extracting("type", "rank", "comment")
-			.containsExactlyInAnyOrder(
-				tuple(FairyType.WIN.getName(), 1, null),
-				tuple(FairyType.WIN.getName(), 2, null),
-				tuple(FairyType.LOSE.getName(), 1, null)
-			);
-	}
-
-	@DisplayName("특정 사용자의 보낸 선물 내역을 조회한다.")
-	@Test
-	void findGiftHistories() {
-		//given
-		Member member = Member.builder()
-			.nickname("테스트유저1")
-			.username("playdot1")
-			.password("123")
-			.socialType(SocialType.KAKAO)
-			.build();
-
-		Member member2 = Member.builder()
-			.nickname("테스트유저2")
-			.username("playdot2")
-			.password("123")
-			.socialType(SocialType.KAKAO)
-			.build();
-
-		memberRepository.saveAll(List.of(member, member2));
-
-		GiftToken giftToken1 = GiftToken.builder()
-			.tokenAmount(1)
-			.giveMember(member2)
-			.takeMember(member)
-			.comment("aaaa")
-			.build();
-
-		GiftToken giftToken2 = GiftToken.builder()
-			.tokenAmount(3)
-			.giveMember(member)
-			.takeMember(member2)
-			.comment("bbbb")
-			.build();
-
-		giftTokenRepository.saveAll(List.of(giftToken1, giftToken2));
-
-		//when
-		Page<MemberResponse.GiftHistoryDTO> result = memberService.findGiftHistories(member.getId(), 0, 15);
-
-		//then
-		assertThat(result.get()).hasSize(1)
-			.extracting("nickname", "token", "comment")
-			.containsExactlyInAnyOrder(
-				tuple("테스트유저2", 3, "bbbb")
-			);
 	}
 
 	@DisplayName("닉네임으로 토큰을 선물한다.")
@@ -450,7 +264,7 @@ class MemberServiceTest {
 		ChatRequestDTO.ChatGiftRequestDTO requestDTO = new ChatRequestDTO.ChatGiftRequestDTO("테스트유저2", 5, "잘쓰세요");
 
 		//when
-		memberService.saveGiftToken(giveMember.getId(), requestDTO);
+		profileService.saveGiftToken(giveMember.getId(), requestDTO);
 
 		//then
 		Member savedGiveMember = memberServiceHelper.getMember(giveMember.getId());
@@ -484,7 +298,7 @@ class MemberServiceTest {
 
 		//when
 		//then
-		assertThatThrownBy(() -> memberService.saveGiftToken(giveMember.getId(), requestDTO))
+		assertThatThrownBy(() -> profileService.saveGiftToken(giveMember.getId(), requestDTO))
 			.isInstanceOf(BusinessException.class)
 			.hasMessage(ErrorCode.INSUFFICIENT_TOKENS.getMessage());
 	}
@@ -508,7 +322,7 @@ class MemberServiceTest {
 
 		//when
 		//then
-		assertThatThrownBy(() -> memberService.saveGiftToken(giveMember.getId(), requestDTO))
+		assertThatThrownBy(() -> profileService.saveGiftToken(giveMember.getId(), requestDTO))
 			.isInstanceOf(BusinessException.class)
 			.hasMessage(ErrorCode.GIFTING_TO_SELF_NOT_ALLOWED.getMessage());
 	}
@@ -538,7 +352,7 @@ class MemberServiceTest {
 		ChatRequestDTO.ChatGiftRequestDTO requestDTO = new ChatRequestDTO.ChatGiftRequestDTO("테스트유저2", 1, "잘쓰세요");
 
 		//when
-		memberService.saveGiftToken(giveMember.getId(), requestDTO);
+		profileService.saveGiftToken(giveMember.getId(), requestDTO);
 
 		//then
 		GiftToken giftToken = giftTokenRepository.findAll().get(0);
@@ -552,5 +366,4 @@ class MemberServiceTest {
 		assertThat(savedTakeMember.getId()).isEqualTo(giveMember.getId());
 		assertThat(savedGiveMember.getId()).isEqualTo(takeMember.getId());
 	}
-
 }
